+++ resources/js/components/alerts/FormError.vue
@@ -1,6 +1,6 @@
 <template>
     <div v-if="validationErrors">
-        <span v-for="(value, key, index) in validationErrors" class="text-error ml-2" :key="index">
+        <span v-for="(value, key, index) in validationErrors" class="text-error px-2" :key="index">
             {{ value }}
         </span>
     </div>