+++ CHANGELOG.md
@@ -6,9 +6,11 @@
 - Added project files to repository
 - Added CHANGELOG.md
 - Added condition on storing of event if only one event per date period is allowed
+- Added date validation to fields from and to in Requests/StoreCalendarEvent
 
 ### Changed
 - Changed store condition of event to if only one event per date period is allowed
 
 ### Removed
 - Removed drawer icon on App.vue
+- Removed CalendarEventController from /v1 folder